move packages and use java8 compliant syntax
