+++ slack-api-model/src/main/java/com/slack/api/util/predicate/FieldPredicate.java
@@ -1,8 +1,8 @@
-package com.slack.api.model.predicate;
+package com.slack.api.util.predicate;
 
 /**
  * A functional interface for defining validation predicates against {@link java.lang.reflect.Field}.  Used by
- * {@link com.slack.api.model.annotation.Required} during object serialization and deserialization to ensure the
+ * {@link com.slack.api.util.annotation.Required} during object serialization and deserialization to ensure the
  * field member is "valid" per the defined predicate.
  */
 @FunctionalInterface
