+++ slack-api-model/src/main/java/com/slack/api/util/predicate/IsNotNullFieldPredicate.java
@@ -1,4 +1,4 @@
-package com.slack.api.model.predicate;
+package com.slack.api.util.predicate;
 
 import java.util.Objects;
